app-next: move core API to frontend-{app,plugin}-api
