+++ packages/frontend-app-api/api-report.md
@@ -3,5 +3,12 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-// (No @packageDocumentation comment for this package)
+/// <reference types="react" />
+
+import { BackstagePlugin } from '@backstage/frontend-plugin-api';
+
+// @public (undocumented)
+export function createApp(options: { plugins: BackstagePlugin[] }): {
+  createRoot(): JSX.Element;
+};
 ```
