+++ packages/frontend-app-api/package.json
@@ -28,5 +28,14 @@
   },
   "files": [
     "dist"
-  ]
+  ],
+  "dependencies": {
+    "@backstage/config": "workspace:^",
+    "@backstage/frontend-plugin-api": "workspace:^",
+    "@backstage/plugin-graphiql": "workspace:^"
+  },
+  "peerDependencies": {
+    "react": "*",
+    "react-router-dom": "*"
+  }
 }
