+++ packages/frontend-app-api/src/index.ts
@@ -13,4 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export {};
+
+/**
+ * TODO
+ *
+ * @packageDocumentation
+ */
+
+export { createApp } from './createApp';
