+++ packages/frontend-plugin-api/package.json
@@ -28,5 +28,11 @@
   },
   "files": [
     "dist"
-  ]
+  ],
+  "peerDependencies": {
+    "react": "^16.13.1 || ^17.0.0"
+  },
+  "dependencies": {
+    "lodash": "^4.17.21"
+  }
 }
